Use category_test for \core_customfield\category_controller test.
Any \core_customfield\api::* tests are moved to api_test.
